Optimal Path Found
Algorithm can find the optimal path but it does not trace the correct path.  Should be an easy fix.  Then change the step and find path buttons to work properly (loop or no loop), change colors, and add some kind of timer on the find path loop so it is easier for the user to see.